CLN/REF: Unify Arithmetic Methods (#27413)
